+++ src/components/DataTable/index.ts
@@ -1,0 +1,9 @@
+import DataTable from './DataTable.vue'
+
+import { App } from 'vue'
+
+DataTable.install = (app: App) => {
+  app.component(DataTable.name, DataTable)
+}
+
+export default DataTable